Pull out read text file.
