+++ src/bundler.js
@@ -56,6 +56,7 @@ var fs = require("fs"),
     _ = require('underscore'),
     collection = require('./collection'),
     cssValidator = require('./css-validator'),
+    readTextFile = require('./read-text-file'),
     compile = require('./compile'),
     minify = require('./minify'),
     urlVersioning = null;
@@ -648,21 +649,4 @@ function compileAsync(options, mode, compileFn /*compileFn(text, textPath, cb(co
 
 function removeCR(text) {
     return text.replace(/\r/g, '');
-}
-
-function stripBOM(content) {
-    // Remove byte order marker. This catches EF BB BF (the UTF-8 BOM)
-    // because the buffer-to-string conversion in `fs.readFileSync()`
-    // translates it to FEFF, the UTF-16 BOM.
-    if (content.charCodeAt(0) === 0xFEFF) {
-        content = content.slice(1);
-    }
-    return content;
-}
-
-function readTextFile(filePath, cb) {
-    fs.readFile(filePath, 'utf-8', function(err, fileContents) {
-        if (err) throw err;
-        cb(stripBOM(fileContents));
-    });
 }
